Refactor code so not everything sits in one file. Adjusting failing tests after is_wizard check removal
